refactor(example): direct flows — drop serviceValue/txBoundary, inline forwards
YAGNI pass: the example is now plain flows, as direct as possible.

- Delete tx.ts (serviceValue store record) and unit.ts (txBoundary unit
  of work). Store operations become flows that dep the database and run
  SQL directly; multi-write ops use a per-op db.transaction. Without a
  per-flow tx boundary, SQL commits inline, so a flow that mutates then
  bumps a signal is naturally post-commit — the old unit-owner/announcer
  split collapses (enqueue is one flow again).
- Inline the single-use forwards: releaseReminder into sendReminder's
  failure path, and deliver into sendReminder — a static tag is not a
  reason for a wrapper flow, so the email step tag now rides the
  client.send.exec({ tags }) call instead of a deliver flow.
- Reuse only where it matters: reads/settle/markReminderSent stay flows
  (exec'd by cli/server/tests as flow refs); everything single-use is
  inlined.

Trade: batch atomicity reverts to per-invoice (each settle its own tx);
serviceValue is no longer demonstrated (traced + role-tag flow remain).
23 example / lint 0 / tc 0; daemon, server, cli verified live against
Postgres, reminder delivered through the inlined traced client.send.
